Fix students not being able to log in if there's no active session by tweaking the Auth service and Student repo logic.
Fix the role gaurd logic that's causing 403 (Forbidden error) when the user actually have the right role to access the resources behind an endpoint.

Sanitized the JSON response being sent to the frontend application.

Adjusted the function that handles CSV upload to return the appropriate number of the records it processed by manually awaiting the ASYNC operation (tuff one this fix wheew)