Added the log out endpoint and the service logic that handles the feature.
Automate Migration in Render by adding migration script in the package.json file.

Configure TypeORM for prod (to auto add new table) using data-source file.

Handle the bug that occurred from throwing exception "403" when student tried to log in by using a return method to provide the appropriate response to the frontend.

Added SSL so Render allows the backend app to connect securely to the DB.